Create the (psuedo) backend to add book to the DOM
We use the `formdata` event since it provides a semantic way to retrieve
the form data without submitting (not submitting form data per the
project specs).

Use form.reset() since it is the semantic way of clearing a form, as
opposed to just setting the fields to the empty string.

Some initial stylings were added to make the testing process less
visually jarring...