Fix cancel button not closing modal & refactor
Refactor to use more semantic interactions between forms and dialogs.
Alternatively using Javascript, can use .preventDefault() to prevent
transmital and call .close() on the dialog to close it, like was done
before.

It is worth noting that the "cancel" button still fires a Submit event
(which is handled with our event handler).