Fix 'Read' button changing row height on Windows
It appears that Ubuntu and Windows both render the checkmark symbol ✓
(U+2713) differently -- Windows' version is greater in height than
the alphanumeric characters (a-z, 0-9) and the cross × character.
This and that each row height is flexible cause toggling a book from
unread to read increase the row height, and toggling from read to unread
decrease the row height; undesired behavior.

Decreasing the font size of the Read column was found to fix this issue.

It's also worth noting the checkmark symbol does not seem to be covered
by the Roboto font family or other major font families, so changing the
font-family attribute did not appear to fix the issue.